Revert "Feat:add redis and postgresql support"
This reverts commit a26719dbaecee0b2629afc411e644032054c6f1e.
