+++ requirements.txt
@@ -10,13 +10,11 @@ motor
 ntgcalls
 pillow
 psutil
-psycopg
 pykeyboard
 py-tgcalls
 git+https://github.com/KurimuzonAkuma/pyrogram@dev
 python-dotenv
 pyyaml
-redis
 requests
 speedtest-cli
 spotipy
